refactor(controller): extract usage-logging.ts (R-util)
