+++ packages/llm-agent-server-libs/src/smart-agent/controller/controller-coordinator-handler.ts
@@ -3,11 +3,9 @@ import {
   externalToolCallId,
   type IEmbedder,
   type IKnowledgeRagHandle,
-  type IRequestLogger,
   type IStageHandler,
   type KnowledgeEntry,
   type KnowledgeEntryMetadata,
-  type LlmComponent,
   type LlmTool,
   type LlmToolCall,
   type LlmUsage,
@@ -46,6 +44,7 @@ import type {
   SessionBundle,
   Step,
 } from './types.js';
+import { makeLogUsage } from './usage-logging.js';
 
 // ---------------------------------------------------------------------------
 // Debug logging — gated behind DEBUG_CONTROLLER (e.g. DEBUG_CONTROLLER=1).
@@ -62,48 +61,8 @@ export type TerminalUsage = LlmUsage & {
   models?: Record<string, ModelUsageEntry>;
 };
 
-/**
- * Build a request-time `logUsage(role, usage)` that writes each subagent call
- * into the per-request `IRequestLogger` (the single aggregator), attributing the
- * role's configured model. The role is explicit at the call site, so the shared
- * planner/finalizer client is attributed correctly. `durationMs: 0` — the seam
- * carries no timing (matches the rag-query precedent).
- */
-export function makeLogUsage(
-  requestLogger: IRequestLogger,
-  requestId: string | undefined,
-  models: {
-    evaluator: string;
-    planner: string;
-    executor: string;
-    reviewer?: string;
-    finalizer?: string;
-  },
-): (role: string, u?: LlmUsage) => void {
-  return (role, u) => {
-    if (!u) return;
-    const model =
-      role === 'finalizer'
-        ? (models.finalizer ?? models.planner)
-        : role === 'reviewer'
-          ? (models.reviewer ?? models.planner)
-          : role === 'embedding'
-            ? 'embedder'
-            : ((models as Record<string, string>)[role] ?? 'unknown');
-    requestLogger.logLlmCall({
-      component: role as LlmComponent,
-      model,
-      promptTokens: u.promptTokens ?? 0,
-      completionTokens: u.completionTokens ?? 0,
-      totalTokens: u.totalTokens ?? 0,
-      durationMs: 0,
-      requestId,
-    });
-    dlog(
-      `tokens ${role}: prompt=${u.promptTokens} completion=${u.completionTokens} total=${u.totalTokens}`,
-    );
-  };
-}
+// Re-exported for import-path stability (moved to ./usage-logging.ts).
+export { makeLogUsage } from './usage-logging.js';
 
 // ---------------------------------------------------------------------------
 // Dep-injection surface
